Move database access into root package and normalize it.
Previously, the web app had direct database code sprinkled throughout it.
Now, all such code should be in files specifically for accessing parts of
the database.

This will make it easier in future access to the database from the web app
(not rewriting code), as well as accessing the program from another entry
point, such as the command line or as a library, without needing to import
the web app itself.